+++ package.json
@@ -14,6 +14,7 @@
     "@emotion/styled": "^11.13.0",
     "@mui/material": "^6.1.3",
     "axios": "^1.7.7",
+    "express": "^4.21.1",
     "lodash": "^4.17.21",
     "msw": "^2.4.10",
     "react": "^18.3.1",
